+++ projects/fhi-angular-treeview-builder/src/public-api.ts
@@ -1,6 +1,0 @@
-/*
- * Public API Surface of fhi-angular-treeview-builder
- */
-
-export * from './lib/fhi-angular-treeview-builder.component';
-export * from './lib/fhi-angular-treeview-builder.module';