+++ src/MOCK_DB_DATA/library-items/prototype-angular-tree/_prototype-angular-tree.data.ts
@@ -1,7 +1,7 @@
 import { LibraryItem } from 'src/app/views/shared/models/library-item.model';
 
-import { TreeviewBuilder } from './prototype-angular-tree';
+import { TreeViewBuilder } from './prototype-angular-tree';
 
 export const PrototypeAngularTreeData: LibraryItem[] = [
-  ...TreeviewBuilder
+  ...TreeViewBuilder
 ];