+++ src/MOCK_DB_DATA/library-items/prototype-angular-tree/_prototype-angular-tree.ids.ts
@@ -1,3 +1,3 @@
 export const PrototypeAngularTreeIds = {
-  TreeviewBuilder: undefined
+  TreeViewBuilder: undefined
 }