+++ src/MOCK_DB_DATA/library-items/prototype-angular-tree/prototype-angular-tree.ts
@@ -2,8 +2,8 @@ import { LibraryItemIds } from '../library-item-ids';
 import { LibraryItem, LibraryItemType } from 'src/app/views/shared/models/library-item.model';
 import { LibraryItemConstants as CONST } from '../library-item-constants';
 
-export const TreeviewBuilder: LibraryItem[] = [{
-  id: LibraryItemIds.TreeviewBuilder,
+export const TreeViewBuilder: LibraryItem[] = [{
+  id: LibraryItemIds.TreeViewBuilder,
   title: 'Treeview Builder',
   type: LibraryItemType.fhiAngular,
   exampleHtml: getExampleHtml(),