+++ src/app/views/shared/dynamic-library-examples/dynamic-library-examples.module.ts
@@ -33,7 +33,7 @@ import { TreeViewCheckboxExampleComponent } from './fhi-angular-examples/tree-vi
 import { PrototypeFormsComponent } from './fhi-prototypes/forms/forms.component';
 import { PrototypePageheaderExampleComponent } from './fhi-prototypes/pageheader/pageheader.component';
 import { TableWithExpandableContentComponent } from './fhi-prototypes/table-with-expandable-content/table-with-expandable-content.component';
-import { AngularTreeComponentExampleComponent } from './fhi-prototypes/angular-tree-component-example/angular-tree-component-example.component';
+import { AngularTreeComponentExampleComponent } from './fhi-prototypes/angular-tree-view-builder-example/angular-tree-view-builder-example.component';
 
 const NGB_EXAMPLES = [
   AccordionExampleComponent,