+++ projects/fhi-angular-tree-view-builder/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 > Mar 27, 2023
 
-* :gift: **New** Create `@folkehelseinstituttet/angular-treeview-builder`
+* :gift: **New** Create `@folkehelseinstituttet/angular-tree-view-builder`
 
 ## Examples